requirements.txt
tabulate==0.9.0
mysql-connector-python==9.1.0